Split out io system calls in the libc backend.
